requirements.txt
faiss-cpu
numpy
pandas
torch
sentence-transformers
transformers
pypdf
python-docx
tqdm
